Fix uge bug in process_triggers that would have resulted in mbf stopping matching triggers after the first iteration of the trigger list; that was supposed to be a while loop
The for loop had the call to get data from the telnet socket inside it, when I meant for it to be inside the while loop but before the trigger processing for loop.
Now it should act as intented; getting data from the socket and then checking every trigger for matches. This is probably the slowest possible way of doing what I want, but I'll improve it later if it's an issue.